+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-17T02:38:11.740Z",
+    "2026-09-17T13:05:36.548Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {