+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1764,7 +1764,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2PKNB60QEMYNS9DDFF3BG1D"),
+  _id: z.string().default("tool_01M2QQJ7DA5GBG7PGRH7HD5X4P"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -1959,7 +1959,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2PKNB60Y59EV2ZYP4QBTKD0"),
+  _id: z.string().default("tool_01M2QQJ7D9EMX2GYZECV0XPQXA"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2054,7 +2054,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2PKNB5Z7EM4WRJ38ZBGVFHF"),
+  _id: z.string().default("tool_01M2QQJ7D8YAFM9EDH3KXKJW0K"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2157,7 +2157,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M2PKNB5Z9JM2XY4ATHAM1CW3"),
+  _id: z.string().default("tool_01M2QQJ7D8GDC2VS1X1EDZKD6P"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),