+++ packages/orq-rc/RELEASES.md
@@ -12449,3 +12449,13 @@ Based on:
 - [typescript v4.16.0-rc.0] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.0 - packages/orq-rc
+
+## 2026-09-17 13:08:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.1] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.1 - packages/orq-rc